Update dependencies for Ubuntu 22.04, drop Ubuntu 18.04
Ubuntu 22.04 is using Python 3.10. We cannot find a set of pinned dependency
versions that works both on Python 3.6 (for Ubuntu 18.04) and on Python 3.10
(for Ubuntu 22.04), so we have to drop support for Ubuntu 18.04. I have
regenerated a set of pinned dependencies on Python 3.8 (Ubuntu 20.04) and
verified that tests pass on Ubuntu 22.04.